scopes are working waitin on joseph to get back to me on specific scope in wiki index

+++ app/controllers/wikis_controller.rb
@@ -1,6 +1,6 @@
 class WikisController < ApplicationController
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki)
   end
 
   def show
@@ -17,7 +17,7 @@ class WikisController < ApplicationController
     @wikis.body = params[:wiki][:body]
     @wikis.private = params[:wiki][:private]
     @wikis.user = current_user
-    
+    authorize @wikis
     if @wikis.save 
       redirect_to @wikis, notice: "New wiki was saved successfully."
     else
@@ -28,6 +28,7 @@ class WikisController < ApplicationController
 
   def edit
     @wikis = Wiki.find(params[:id])
+    authorize @wikis
   end
   
   def update
@@ -35,7 +36,7 @@ class WikisController < ApplicationController
     @wikis.title = params[:wiki][:title]
     @wikis.body = params[:wiki][:body]
     @wikis.private = params[:wiki][:private]
-    
+    authorize @wikis
     if @wikis.save
       redirect_to @wikis
     else
@@ -46,7 +47,7 @@ class WikisController < ApplicationController
 
   def destroy
     @wikis = Wiki.find(params[:id])
-    
+    authorize @wikis
     if @wikis.destroy
       redirect_to wikis_path
     else
